[#72] processors.py -> styler.py / explainer.py / translator.py
