+++ politely/__init__.py
@@ -1,2 +1,4 @@
-from politely.processors import Styler, Translator, Explainer
+from politely.styler import Styler
+from politely.explainer import Explainer
+from politely.translator import Translator
